Move and rename TinkerTabsScreen to TinkerTabsWidget
